requirements/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements/requirements.in
#
anyio==3.5.0
    # via starlette
asgiref==3.5.0
    # via uvicorn
bcrypt==3.2.0
    # via passlib
cffi==1.15.0
    # via
    #   bcrypt
    #   cryptography
click==8.0.4
    # via uvicorn
colorama==0.4.4
    # via click
cryptography==36.0.1
    # via python-jose
dnspython==2.2.0
    # via pymongo
ecdsa==0.17.0
    # via python-jose
fastapi==0.75.0
    # via -r requirements/requirements.in
h11==0.13.0
    # via uvicorn
idna==3.3
    # via anyio
motor==2.5.1
    # via -r requirements/requirements.in
passlib[bcrypt]==1.7.4
    # via -r requirements/requirements.in
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycparser==2.21
    # via cffi
pydantic==1.9.0
    # via fastapi
pyhumps==3.5.3
    # via -r requirements/requirements.in
pymongo[srv]==3.12.3
    # via
    #   -r requirements/requirements.in
    #   motor
python-jose[cryptography]==3.3.0
    # via -r requirements/requirements.in
python-multipart==0.0.5
    # via -r requirements/requirements.in
rsa==4.8
    # via python-jose
six==1.16.0
    # via
    #   bcrypt
    #   ecdsa
    #   python-multipart
sniffio==1.2.0
    # via anyio
starlette==0.17.1
    # via fastapi
typing-extensions==4.1.1
    # via pydantic
uvicorn==0.17.5
    # via -r requirements/requirements.in
